+++ aldrin/CHANGELOG.md
@@ -31,6 +31,8 @@ This version still connects to brokers using protocol version 14.
 - Add `BusListener::new()`.
 - Add `Discoverer::client()`.
 - Add `Lifetime::client()`.
+- Add `client()` getters to `PendingSender`, `PendingReceiver`, `UnclaimedSender` and
+  `UnclaimedReceiver`.
 
 ### Fixed